+++ inst/examples/linear_algebra_stats.R
@@ -6,6 +6,6 @@ solve(vc)
 cinv(vc) # faster than `solve()`, but only works on positive definite matrices
 all(round(solve(vc), 6) == round(cinv(vc), 6)) # they're the same
 
-sd_gauss_lc(X, vc)
+sd_lc(X, vc)